Migrate document editor to Lexical 0.44 Extension API
Switch the document editor from LexicalComposer + plugin-list to
LexicalExtensionComposer + a single AppExtension built via
defineExtension. Most plugin imports become extension dependencies:
RichTextExtension, HistoryExtension (skipped in collab mode where Yjs
owns undo), LinkExtension/AutoLinkExtension/ClickableLinkExtension,
ListExtension/CheckListExtension, HorizontalRuleExtension,
ClearEditorExtension, AutoFocusExtension,
SelectionAlwaysOnDisplayExtension, DecoratorTextExtension,
HashtagExtension, plus the new CodeExtension + CodePrismExtension that
replace the custom code-highlight-plugin and clear the dev-mode
deprecation warning around CodeNode.

Custom extensions (frontend/src/components/ui/editor/extensions/):
markdown-shortcuts, emojis, images, keywords, layout, heading-anchor,
list-max-indent-level. Toolbar / floating / picker / mention / wikilink
/ embed plugins stay as JSX children of LexicalExtensionComposer, since
they render React DOM and have no clean extension form (matching the
upstream shadcn-editor pattern).

Two correctness notes:

- The AppExtension is stabilized via useRef snapshots so the editor is
  not destroyed mid-session when the parent loads editorSerializedState
  asynchronously. LexicalExtensionComposer recreates the editor
  whenever its `extension` prop reference changes, which would
  otherwise wipe out an active Yjs sync after the document fetched.

- useCollaboration's providerFactory now always seeds the incoming
  yjsDocMap with the existing provider's Y.Doc on the
  reuse-existing-provider path. Previously the early return skipped
  yjsDocMap.set(); under LexicalExtensionComposer the docMap can be a
  fresh instance per mount, so CollaborationPlugin's
  setDoc(yjsDocMap.get(id)) returned undefined and createBinding threw
  "doc is null or undefined".

No user-visible behavior change.
