+++ frontend/package.json
@@ -37,8 +37,11 @@
     "@excalidraw/excalidraw": "^0.18.1",
     "@icons-pack/react-simple-icons": "^13.13.0",
     "@lexical/code": "^0.44.0",
+    "@lexical/code-prism": "^0.44.0",
+    "@lexical/extension": "^0.44.0",
     "@lexical/file": "^0.44.0",
     "@lexical/hashtag": "^0.44.0",
+    "@lexical/history": "^0.44.0",
     "@lexical/html": "^0.44.0",
     "@lexical/link": "^0.44.0",
     "@lexical/list": "^0.44.0",
