+++ frontend/src/components/ui/editor/extensions/images-extension.ts
@@ -1,0 +1,185 @@
+import { defineExtension } from "lexical";
+import { $wrapNodeInElement, mergeRegister } from "@lexical/utils";
+import {
+  $createParagraphNode,
+  $createRangeSelection,
+  $getSelection,
+  $insertNodes,
+  $isNodeSelection,
+  $isRootOrShadowRoot,
+  $setSelection,
+  COMMAND_PRIORITY_EDITOR,
+  COMMAND_PRIORITY_HIGH,
+  COMMAND_PRIORITY_LOW,
+  DRAGOVER_COMMAND,
+  DRAGSTART_COMMAND,
+  DROP_COMMAND,
+  type LexicalEditor,
+} from "lexical";
+
+import { $createImageNode, $isImageNode, ImageNode } from "@/components/ui/editor/nodes/image-node";
+import { CAN_USE_DOM } from "@/components/ui/editor/shared/can-use-dom";
+import {
+  INSERT_IMAGE_COMMAND,
+  type InsertImagePayload,
+} from "@/components/ui/editor/plugins/images-plugin";
+
+const getDOMSelection = (targetWindow: Window | null): Selection | null =>
+  CAN_USE_DOM ? (targetWindow || window).getSelection() : null;
+
+function $getImageNodeInSelection(): ImageNode | null {
+  const selection = $getSelection();
+  if (!$isNodeSelection(selection)) {
+    return null;
+  }
+  const nodes = selection.getNodes();
+  const node = nodes[0];
+  return $isImageNode(node) ? node : null;
+}
+
+function getDragImageData(event: DragEvent): null | InsertImagePayload {
+  const dragData = event.dataTransfer?.getData("application/x-lexical-drag");
+  if (!dragData) {
+    return null;
+  }
+  const { type, data } = JSON.parse(dragData);
+  if (type !== "image") {
+    return null;
+  }
+  return data;
+}
+
+function canDropImage(event: DragEvent): boolean {
+  const target = event.target;
+  return !!(
+    target &&
+    target instanceof HTMLElement &&
+    !target.closest("code, span.editor-image") &&
+    target.parentElement &&
+    target.parentElement.closest("div.ContentEditable__root")
+  );
+}
+
+function getDragSelection(event: DragEvent): Range | null | undefined {
+  let range;
+  const target = event.target as null | Element | Document;
+  const targetWindow =
+    target == null
+      ? null
+      : target.nodeType === 9
+        ? (target as Document).defaultView
+        : (target as Element).ownerDocument.defaultView;
+  const domSelection = getDOMSelection(targetWindow);
+  if (document.caretRangeFromPoint) {
+    range = document.caretRangeFromPoint(event.clientX, event.clientY);
+  } else if (event.rangeParent && domSelection !== null) {
+    domSelection.collapse(event.rangeParent, event.rangeOffset || 0);
+    range = domSelection.getRangeAt(0);
+  } else {
+    throw Error(`Cannot get the selection when dragging`);
+  }
+  return range;
+}
+
+function $onDragStart(event: DragEvent): boolean {
+  const node = $getImageNodeInSelection();
+  if (!node) {
+    return false;
+  }
+  const dataTransfer = event.dataTransfer;
+  if (!dataTransfer) {
+    return false;
+  }
+  const TRANSPARENT_IMAGE =
+    "data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7";
+  const img = document.createElement("img");
+  img.src = TRANSPARENT_IMAGE;
+  dataTransfer.setData("text/plain", "_");
+  dataTransfer.setDragImage(img, 0, 0);
+  dataTransfer.setData(
+    "application/x-lexical-drag",
+    JSON.stringify({
+      data: {
+        altText: node.__altText,
+        caption: node.__caption,
+        height: node.__height,
+        key: node.getKey(),
+        maxWidth: node.__maxWidth,
+        showCaption: node.__showCaption,
+        src: node.__src,
+        width: node.__width,
+      },
+      type: "image",
+    })
+  );
+  return true;
+}
+
+function $onDragover(event: DragEvent): boolean {
+  const node = $getImageNodeInSelection();
+  if (!node) {
+    return false;
+  }
+  if (!canDropImage(event)) {
+    event.preventDefault();
+  }
+  return true;
+}
+
+function $onDrop(event: DragEvent, editor: LexicalEditor): boolean {
+  const node = $getImageNodeInSelection();
+  if (!node) {
+    return false;
+  }
+  const data = getDragImageData(event);
+  if (!data) {
+    return false;
+  }
+  event.preventDefault();
+  if (canDropImage(event)) {
+    const range = getDragSelection(event);
+    node.remove();
+    const rangeSelection = $createRangeSelection();
+    if (range !== null && range !== undefined) {
+      rangeSelection.applyDOMRange(range);
+    }
+    $setSelection(rangeSelection);
+    editor.dispatchCommand(INSERT_IMAGE_COMMAND, data);
+  }
+  return true;
+}
+
+export const ImagesExtension = defineExtension({
+  name: "@initiative/images",
+  nodes: [ImageNode],
+  register: (editor) =>
+    mergeRegister(
+      editor.registerCommand<InsertImagePayload>(
+        INSERT_IMAGE_COMMAND,
+        (payload) => {
+          const imageNode = $createImageNode(payload);
+          $insertNodes([imageNode]);
+          if ($isRootOrShadowRoot(imageNode.getParentOrThrow())) {
+            $wrapNodeInElement(imageNode, $createParagraphNode).selectEnd();
+          }
+          return true;
+        },
+        COMMAND_PRIORITY_EDITOR
+      ),
+      editor.registerCommand<DragEvent>(
+        DRAGSTART_COMMAND,
+        (event) => $onDragStart(event),
+        COMMAND_PRIORITY_HIGH
+      ),
+      editor.registerCommand<DragEvent>(
+        DRAGOVER_COMMAND,
+        (event) => $onDragover(event),
+        COMMAND_PRIORITY_LOW
+      ),
+      editor.registerCommand<DragEvent>(
+        DROP_COMMAND,
+        (event) => $onDrop(event, editor),
+        COMMAND_PRIORITY_HIGH
+      )
+    ),
+});
