--- frontend/src/components/ui/editor/plugins/code-highlight-plugin.tsx
@@ -1,13 +1,0 @@
-import { JSX, useEffect } from "react";
-import { registerCodeHighlighting } from "@lexical/code";
-import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-
-export function CodeHighlightPlugin(): JSX.Element | null {
-  const [editor] = useLexicalComposerContext();
-
-  useEffect(() => {
-    return registerCodeHighlighting(editor);
-  }, [editor]);
-
-  return null;
-}
